Fixed click-run bug
In the earlier commit, the ball used to spawn outside the wall if click there, although it would automatically teleport inside the box as an unexpected result of some code written earlier.

This commit makes it such that the click-run function only runs when the mouse is clicked inside the wall boundaries.